+++ app/src/main/java/com/gangwonhyuil/gangwonhyuil/ui/community/screen/addPlace/AddPlaceActivity.kt
@@ -1,4 +1,4 @@
-package com.gangwonhyuil.gangwonhyuil.ui.community.addPost
+package com.gangwonhyuil.gangwonhyuil.ui.community.screen.addPlace
 
 import android.os.Bundle
 import androidx.activity.enableEdgeToEdge
@@ -7,11 +7,11 @@ import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
 import com.gangwonhyuil.gangwonhyuil.R
 
-class AddPostActivity : AppCompatActivity() {
+class AddPlaceActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
-        setContentView(R.layout.activity_add_post)
+        setContentView(R.layout.activity_add_place)
         ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main)) { v, insets ->
             val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
             v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
