+++ app/src/main/res/layout/activity_add_place.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".ui.community.addPlace.AddPlaceActivity">
+    tools:context=".ui.community.screen.addPlace.AddPlaceActivity">
 
     <com.google.android.material.appbar.MaterialToolbar
         android:layout_width="match_parent"
