+++ app/src/main/res/layout/activity_add_post.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".ui.community.addPost.AddPostActivity">
+    tools:context=".ui.community.screen.addPost.AddPostActivity">
 
     <com.google.android.material.appbar.MaterialToolbar
         android:layout_width="match_parent"
