+++ app/src/main/res/layout/fragment_community.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".ui.community.CommunityFragment">
+    tools:context=".ui.community.screen.community.CommunityFragment">
 
     <com.google.android.material.appbar.MaterialToolbar
         android:id="@+id/topAppBar"
